Update advanced_interface.js
change in line 1047 from "clipDEM.toByte()" to "clipDEM.toFloat()". Now, there should not be issues with raster format when exporting the DEM.

+++ app_code/advanced_interface.js
@@ -1044,7 +1044,7 @@ function doMosaic() {
   var exportDEM = function(){
     print('Exporting DEM. Go to the Task tab and press Run...');
     Export.image.toDrive({
-      image: clipDEM.toByte(),
+      image: clipDEM.toFloat(),
       description: 'DEM_Wetlands',
       folder: 'Wetlands',
       scale: 2,
